+++ common/src/main/java/com/wastedrivinggroup/consumer/netty/Constant.java
@@ -1,0 +1,16 @@
+package com.wastedrivinggroup.consumer.netty;
+
+/**
+ * @author chen
+ * @date 2021/6/15
+ **/
+public interface Constant {
+
+	interface ErrCode {
+		/**
+		 * 服务未找到
+		 */
+		int SERVICE_NOT_FOUND = -1;
+	}
+
+}
