+++ common/src/main/java/com/wastedrivinggroup/consumer/netty/client/AbstractNettyClientBootstrap.java
@@ -1,0 +1,18 @@
+package com.wastedrivinggroup.consumer.netty.client;
+
+import com.wastedrivinggroup.consumer.netty.RpcBootstrap;
+
+/**
+ * 客户端(消费者端)的启动类
+ *
+ * @author chen
+ * @date 2021/5/2
+ **/
+public class AbstractNettyClientBootstrap implements RpcBootstrap {
+
+
+    @Override
+    public void start() throws InterruptedException {
+
+    }
+}
