+++ common/src/main/java/com/wastedrivinggroup/consumer/netty/exception/ServiceNotFoundException.java
@@ -1,0 +1,14 @@
+package com.wastedrivinggroup.consumer.netty.exception;
+
+import com.wastedrivinggroup.consumer.netty.Constant;
+
+/**
+ * @author chen
+ * @date 2021/6/15
+ **/
+public class ServiceNotFoundException extends RpcException {
+
+	public ServiceNotFoundException(String serviceName) {
+		super(Constant.ErrCode.SERVICE_NOT_FOUND, serviceName + "服务未找到");
+	}
+}
