--- common/src/main/java/netty/Constant.java
@@ -1,16 +1,0 @@
-package netty;
-
-/**
- * @author chen
- * @date 2021/6/15
- **/
-public interface Constant {
-
-	interface ErrCode {
-		/**
-		 * 服务未找到
-		 */
-		int SERVICE_NOT_FOUND = -1;
-	}
-
-}
