--- common/src/main/java/netty/client/AbstractNettyClientBootstrap.java
@@ -1,18 +1,0 @@
-package netty.client;
-
-import netty.RpcBootstrap;
-
-/**
- * 客户端(消费者端)的启动类
- *
- * @author chen
- * @date 2021/5/2
- **/
-public class AbstractNettyClientBootstrap implements RpcBootstrap {
-
-
-    @Override
-    public void start() throws InterruptedException {
-
-    }
-}
