--- common/src/main/java/netty/exception/ServiceNotFoundException.java
@@ -1,14 +1,0 @@
-package netty.exception;
-
-import netty.Constant;
-
-/**
- * @author chen
- * @date 2021/6/15
- **/
-public class ServiceNotFoundException extends RpcException {
-
-	public ServiceNotFoundException(String serviceName) {
-		super(Constant.ErrCode.SERVICE_NOT_FOUND, serviceName + "服务未找到");
-	}
-}
