--- common/src/main/java/netty/proto/demo/InvokeRespProto.java
@@ -1,9 +1,0 @@
-package netty.proto.demo;
-
-/**
- * @author chen
- * @date 2021/6/16
- **/
-public class InvokeRespProto {
-	String ret;
-}
