--- common/src/main/java/netty/proto/exception/BadProtoFormatException.java
@@ -1,11 +1,0 @@
-package netty.proto.exception;
-
-/**
- * @author chen
- * @date 2021/5/4
- **/
-public class BadProtoFormatException extends RuntimeException {
-    public BadProtoFormatException(String message) {
-        super(message);
-    }
-}
