+++ consumer/src/main/java/Consumer.java
@@ -1,4 +1,6 @@
-import netty.ConsumerBootstrap;
+import com.wastedrivinggroup.consumer.netty.ConsumerBootstrap;
+import com.wastedrivinggroup.consumer.rpc.InvokeProxyHandler;
+import com.wastedrivinggroup.consumer.service.TestService;
 
 /**
  * @author chen
@@ -8,5 +10,7 @@ public class Consumer {
 	public static void main(String[] args) throws InterruptedException {
 		ConsumerBootstrap clent = new ConsumerBootstrap();
 		clent.start();
+		final TestService proxy = InvokeProxyHandler.createProxy(TestService.class);
+		proxy.incr("Hello");
 	}
 }
