+++ consumer/src/main/java/com/wastedrivinggroup/consumer/service/TestService.java
@@ -1,0 +1,9 @@
+package com.wastedrivinggroup.consumer.service;
+
+/**
+ * @author chen
+ * @date 2021/6/16
+ **/
+public interface TestService {
+	int incr(String name);
+}
