+++ pom.xml
@@ -19,12 +19,12 @@
         <java.version>1.8</java.version>
         <lombok.version>1.18.20</lombok.version>
         <netty.version>4.1.63.Final</netty.version>
+        <com.wastedrivinggroup.consumer.netty.version>4.1.63.Final</com.wastedrivinggroup.consumer.netty.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <gson.version>2.8.5</gson.version>
         <commons-io.version>2.8.0</commons-io.version>
         <commons-lang3.version>3.12.0</commons-lang3.version>
         <guava.version>30.1.1-jre</guava.version>
-        <slf4j-api.version>1.7.21</slf4j-api.version>
     </properties>
 
     <!-- dependencyManagement 管理依赖的版本号 -->
@@ -54,14 +54,7 @@
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
-            <version>${slf4j-api.version}</version>
-        </dependency>
-        <!--         https://mvnrepository.com/artifact/org.slf4j/slf4j-simple-->
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-simple</artifactId>
-            <version>${slf4j-api.version}</version>
-            <type>jar</type>
+            <version>1.7.26</version>
         </dependency>
         <!-- https://mvnrepository.com/artifact/ch.qos.logback/logback-core -->
         <dependency>
@@ -74,7 +67,6 @@
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
             <version>1.2.3</version>
-            <scope>test</scope>
         </dependency>
         <!-- https://mvnrepository.com/artifact/org.projectlombok/lombok -->
         <dependency>
