+++ provider/src/main/java/Provider.java
@@ -1,4 +1,7 @@
-import netty.ProviderBootstrap;
+import com.wastedrivinggroup.provider.netty.ProviderBootstrap;
+import com.wastedrivinggroup.provider.service.ServiceHolder;
+import com.wastedrivinggroup.provider.service.TestService;
+import lombok.extern.slf4j.Slf4j;
 
 /**
  * 启动主类
@@ -6,9 +9,11 @@ import netty.ProviderBootstrap;
  * @author chen
  * @date 2021/5/2
  **/
+@Slf4j
 public class Provider {
 
 	public static void main(String[] args) throws InterruptedException {
+		ServiceHolder.getInstance().loadService(TestService.class);
 		ProviderBootstrap server = new ProviderBootstrap();
 		server.start();
 	}
